final commit

+++ app/src/main/java/com/example/bookingbook/ForcedTerminationService.java
@@ -18,11 +18,15 @@ public class ForcedTerminationService extends Service {
 
     @Override
     public void onTaskRemoved(Intent rootIntent) { //핸들링 하는 부분
-        SharedPreferences pref = getSharedPreferences("login", MODE_PRIVATE);
-        SharedPreferences.Editor editor = pref.edit();
-        editor.clear();
-        editor.commit();
-        Toast.makeText(this, "로그아웃되었습니다.", Toast.LENGTH_SHORT).show();
-        stopSelf(); //서비스 종료
+        SharedPreferences sharedPreferences = getSharedPreferences("login", MODE_PRIVATE);    // test 이름의 기본모드 설정, 만약 test key값이 있다면 해당 값을 불러옴.
+        String name = sharedPreferences.getString("login","noname");
+        if(!name.equals("noname"))
+        {
+            SharedPreferences pref = getSharedPreferences("login", MODE_PRIVATE);
+            SharedPreferences.Editor editor = pref.edit();
+            editor.clear();
+            editor.commit();
+            Toast.makeText(this, "로그아웃되었습니다.", Toast.LENGTH_SHORT).show();
+        }
     }
 }
